fix: 审计校验按 bigint 读 status_code，哈希覆盖 jsonb 列字节
operation_logs.status_code 是 bigint，verify 绑成 i32 会整表读失败。
metadata 的哈希原先用 serde 紧凑 JSON，对不上 Postgres jsonb::text。

+++ crates/gw-panel/src/ops/audit_log.rs
@@ -64,7 +64,8 @@ struct AuditRow {
     target: Option<String>,
     method: Option<String>,
     path: Option<String>,
-    status_code: Option<i32>,
+    /// 列是 `bigint`。绑成 `i32` 时 verify 整表读失败。
+    status_code: Option<i64>,
     ip_address: Option<String>,
     request_id: Option<String>,
     metadata: Option<String>,
